refactor(model): move _get_dataset_execution to DerivaMLBagView (Phase 2 §4.5)
The method joins through bag_rids + a Dataset_Version row lookup
— both deriva-ml domain knowledge, not generic bag mechanics.
Per ADR-0006's three-class consumer layer (BagDatabase generic →
DatabaseModel deriva-ml-extended → DerivaMLBagView domain-shaped),
the method belongs in the view, not the database model.

* Moved from model/database.py to model/deriva_ml_bag_view.py.
* Updated two view-internal callers in find_datasets +
  lookup_dataset to ``self._get_dataset_execution(...)``.
* Updated DatasetBag.__init__ caller from
  ``self.model._get_dataset_execution(...)`` to
  ``self._catalog._get_dataset_execution(...)`` — DatasetBag.model
  is the DatabaseModel (where the method used to live);
  DatasetBag._catalog is the view (where it lives now).
* Hoisted the local ``from sqlalchemy import select`` /
  ``from sqlalchemy.orm import Session`` imports in the view to
  module scope (lookup_term already needed them after Step 2;
  now _get_dataset_execution does too).
* ruff swept the now-unused ``Any`` and ``Session`` imports from
  database.py.
* Updated database.py module docstring to point at the new home.

No behaviour change. Layering is honest.
